implement prom suggestions
