+++ boards/DCM/Src/App/states/App_FaultState.c
@@ -1,7 +1,6 @@
 #include "states/App_SharedStates.h"
 #include "states/App_FaultState.h"
 #include "states/App_InitState.h"
-#include "App_SendNonPeriodicCanSignals.h"
 
 #include "App_SharedMacros.h"
 
@@ -30,12 +29,15 @@ static void FaultStateRunOnEntry(struct StateMachine *const state_machine)
 static void FaultStateRunOnTick1Hz(struct StateMachine *const state_machine)
 {
     struct DcmWorld *world = App_SharedStateMachine_GetWorld(state_machine);
+    struct DcmCanRxInterface *can_rx = App_DcmWorld_GetCanRx(world);
 
     // Clear inverter fault if requested by a PCAN node
-    App_SendNonPeriodicCanSignals_ClearInverterFaults(world);
+    App_SharedStates_HandleClearInvFaultsCmd(
+        App_DcmWorld_GetCanTx(world), can_rx);
 
     // Open or close the inverter LV switches if requested by a PCAN node
-    App_SharedStates_ConfigInverterSwitches(world);
+    App_SharedStates_ConfigInverterSwitches(
+        can_rx, App_DcmWorld_GetInverterSwitches(world));
 
     App_SharedStatesRunOnTick1Hz(state_machine);
 }
