+++ boards/DCM/Src/App/states/App_InitState.c
@@ -2,7 +2,6 @@
 #include "states/App_InitState.h"
 #include "states/App_DriveState.h"
 #include "states/App_FaultState.h"
-#include "App_SendNonPeriodicCanSignals.h"
 
 #include "App_SharedMacros.h"
 
@@ -44,12 +43,15 @@ static void InitStateRunOnEntry(struct StateMachine *const state_machine)
 static void InitStateRunOnTick1Hz(struct StateMachine *const state_machine)
 {
     struct DcmWorld *world = App_SharedStateMachine_GetWorld(state_machine);
+    struct DcmCanRxInterface *can_rx = App_DcmWorld_GetCanRx(world);
 
     // Clear inverter fault if requested by a PCAN node
-    App_SendNonPeriodicCanSignals_ClearInverterFaults(world);
+    App_SharedStates_HandleClearInvFaultsCmd(
+        App_DcmWorld_GetCanTx(world), can_rx);
 
     // Open or close the inverter LV switches if requested by a PCAN node
-    App_SharedStates_ConfigInverterSwitches(world);
+    App_SharedStates_ConfigInverterSwitches(
+        can_rx, App_DcmWorld_GetInverterSwitches(world));
 
     App_SharedStatesRunOnTick1Hz(state_machine);
 }
