+++ scripts/codegen/CAN/App_CanMsgs.dbc
@@ -582,10 +582,8 @@ SG_ Clear_Left_Fault : 0|1@1+ (1,0) [0|1] "" DCM
 SG_ Clear_Right_Fault : 1|1@1+ (1,0) [0|1] "" DCM
 
 BO_ 601 INV_LOAD_SWITCHES_CMD: 1 DEBUG
-SG_ Close_Left_Switch : 0|1@1+ (1,0) [0|1] "" DCM
-SG_ Close_Right_Switch : 1|1@1+ (1,0) [0|1] "" DCM
-SG_ Open_Left_Switch : 2|1@1+ (1,0) [0|1] "" DCM
-SG_ Open_Right_Switch : 3|1@1+ (1,0) [0|1] "" DCM
+SG_ Config_Left_Switch : 0|1@1+ (1,0) [0|1] "" DCM
+SG_ Config_Right_Switch : 1|1@1+ (1,0) [0|1] "" DCM
 
 BA_DEF_  "BusType" STRING ;
 BA_DEF_ BO_  "GenMsgCycleTime" INT 0 65535;
@@ -695,7 +693,7 @@ BA_ "GenMsgCycleTime" BO_ 509 1000;
 BA_ "GenMsgCycleTime" BO_ 510 1000;
 
 BA_ "GenMsgCycleTime" BO_ 600 0;
-BA_ "GenMsgCycleTime" BO_ 601 0;
+BA_ "GenMsgCycleTime" BO_ 601 1000;
 
 SIG_VALTYPE_ 106 Duty_Cycle : 1;
 SIG_VALTYPE_ 106 Frequency : 1;
@@ -781,9 +779,7 @@ VAL_ 506 Drive_Mode 0 "DRIVE_MODE_1" 1 "DRIVE_MODE_2" 2 "DRIVE_MODE_3" 3 "DRIVE_
 VAL_ 507 Start_Switch 0 "OFF" 1 "ON";
 VAL_ 507 Traction_Control_Switch 0 "OFF" 1 "ON";
 VAL_ 507 Torque_Vectoring_Switch  0 "OFF" 1 "ON";
-VAL_ 600 Clear_Left_Fault 0 "NO" 1 "YES";
-VAL_ 600 Clear_Right_Fault 0 "NO" 1 "YES";
-VAL_ 601 Close_Left_Switch 0 "NO" 1 "YES";
-VAL_ 601 Close_Right_Switch 0 "NO" 1 "YES";
-VAL_ 601 Open_Left_Switch 0 "NO" 1 "YES";
-VAL_ 601 Open_Right_Switch 0 "NO" 1 "YES";
+VAL_ 600 Clear_Left_Fault 0 "FALSE" 1 "TRUE";
+VAL_ 600 Clear_Right_Fault 0 "FALSE" 1 "TRUE";
+VAL_ 601 Config_Left_Switch 0 "OFF" 1 "ON";
+VAL_ 601 Config_Right_Switch 0 "OFF" 1 "ON";
